fix(diff): keep inline editors alive across a comments/drafts refresh
The prior fix (6be1c60) gave the inline comment zones a reconcile-in-place
controller, but two gaps remained that still interrupted an open inline
editor on every poll:

1. useDiffComments called setComments() unconditionally, so an unchanged
   poll still produced a fresh comments array. That churned the
   mentionCandidates memo, whose identity change re-ran useDraftZones. Bail
   on structural equality (reuse sameCommentList, now extracted to a shared
   commentEquality module) so an unchanged poll keeps the old reference and
   downstream memos stay stable — mirroring what the activity view already
   does.

2. useDraftZones used the one-shot mountInlineZones, so a real drafts/
   comments change tore down and rebuilt every DraftZone root, dropping an
   open draft editor's edit mode / focus / caret. Give it the same split
   structural/content effect + persistent createInlineZones controller as
   useCommentZones, reconciling draft zones by (side, line) key. A surviving
   draft re-renders in place (DraftZoneList keys by draft id); only a removed
   draft (deleted / published / file switch) unmounts and runs its cancel
   cleanup.

The now-unused one-shot mountInlineZones wrapper is removed. The activity
timeline already bailed via sameCommentList and needs no change.
